requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-docx==0.8.11
email-validator==2.1.0
pydantic==2.5.0
psycopg2-binary==2.9.9
bcrypt==4.0.1